+++ WebAppFAM/Pages/Trips/TripMain.cshtml.cs
@@ -7,9 +7,11 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
 using WebAppFAM.Models;
 using WebAppFAM;
+using Microsoft.AspNetCore.Authorization;
 
 namespace WebAppFAM.Pages.Trips
 {
+    [Authorize]
     public class TripMain : PageModel
     {
         private readonly WebAppFAM.Models.WebAppFAMContext _context;
@@ -99,7 +101,7 @@ namespace WebAppFAM.Pages.Trips
 
         public IActionResult OnDeleteDelete([FromBody] Location obj)
         {
-            if (obj != null)
+            if (obj != null && HttpContext.User.IsInRole("Admin"))
             {
                 _context.Locations.Remove(obj);
                 _context.SaveChanges();
